+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 # Change Log
 
-All notable changes to the "env-vault" extension will be documented in this file.
+All notable changes to the "dot-vault" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+
+## [0.4.0] - 2025-05-06
+
+### Changed
+
+- Renamed project from "Env Vault" to "DotVault" for better branding
+- Updated all references in code and documentation
+- Changed package name from "env-vault" to "dot-vault"
+- Updated GitHub repository references
+- Modified command categories from "EnvVault" to "DotVault"
+
+### Added
+
+- Added MIT LICENSE.md file
+
 
 ## [0.3.2] - 2025-05-06
 
@@ -55,7 +70,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [0.1.4] - Initial release
 
-- Initial version of Env Vault extension
+- Initial version of DotVault extension
 - Secure encryption and storage of .env files in Supabase
 - GitHub authentication via Supabase
 - Basic commands for syncing and restoring .env files